requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5
aiofiles==24.1.0
websockets==14.2

# Task queue
celery[redis]==5.4.0
redis==5.2.1

# Database
sqlalchemy==2.0.36
alembic==1.14.1
aiosqlite==0.20.0

# Pydantic / Config
pydantic==2.10.4
pydantic-settings==2.7.1

# AI/ML
scenedetect[opencv]==0.6.5
open-clip-torch==2.29.0
ultralytics==8.3.57
openai-whisper==20240930
opencv-python-headless==4.10.0.84
torch==2.5.1
torchvision==0.20.1
Pillow==11.1.0

# Export
lxml==5.3.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1

# Optional Enhancement
# demucs==4.0.1
# realesrgan @ git+https://github.com/xinntao/Real-ESRGAN.git
